+++ backend/Hubla.Sales.Application/Features/GetSales/UseCase/GetSalesOutput.cs
@@ -6,12 +6,12 @@ namespace Hubla.Sales.Application.Features.GetSales.UseCase
 {
     public class GetSalesOutput : SaleOutputBase
     {
-        private GetSalesOutput(int id, SaleType saleType, DateTime date, string description, double value, SellerOutputBase seller)
+        private GetSalesOutput(int id, SaleType saleType, DateTime date, string description, decimal value, SellerOutputBase seller)
             : base(id, saleType, date, description, value, seller)
         {
         }
 
-        public static GetSalesOutput Create(int id, SaleType saleType, DateTime date, string description, double value, SellerOutputBase seller) =>
+        public static GetSalesOutput Create(int id, SaleType saleType, DateTime date, string description, decimal value, SellerOutputBase seller) =>
             new(id, saleType, date, description, value, seller);
     }
